requirements.txt
aiogram==3.*
sqlalchemy
aiosqlite
python-dotenv